+++ ch32v307/src/step.cpp
@@ -60,7 +60,7 @@ void StepY::handler() {
         break;
     case START_MOVING:
         // if (currentValue < stopValue) {
-        if (pwm.counter < stopValue * coeff) {
+        if (pwm.counter < (uint32_t)((float)stopValue * coeff)) {
             currentState = MOVING_PLUS;
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::START_Y;
@@ -70,7 +70,7 @@ void StepY::handler() {
             pwm.slow();
             // TODO: block all
             //} else if (currentValue > stopValue) {
-        } else if (pwm.counter > stopValue * coeff) {
+        } else if (pwm.counter > (uint32_t)((float)stopValue * coeff)) {
             currentState = MOVING_MINUS;
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::START_Y;
@@ -88,11 +88,11 @@ void StepY::handler() {
         break;
     case MOVING_PLUS:
         // if (currentValue >= stopValue) {
-        if (pwm.counter >= stopValue * coeff) {
+        if (pwm.counter >= (uint32_t)((float)stopValue * coeff)) {
             stop();
             currentState = STOP;
             // send signal about final place
-            currentValue = pwm.counter / coeff;
+            currentValue = (uint32_t)((float)pwm.counter / coeff);
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::STOP_Y;
             Udp::pThis->reply.val = currentValue;
@@ -119,11 +119,11 @@ void StepY::handler() {
         break;
     case MOVING_MINUS:
         // if (currentValue <= stopValue) {
-        if (pwm.counter <= stopValue * coeff) {
+        if (pwm.counter <= (uint32_t)((float)stopValue * coeff)) {
             stop();
             currentState = STOP;
             // TODO: send signal about final place
-            currentValue = pwm.counter / coeff;
+            currentValue = (uint32_t)((float)pwm.counter / coeff);
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::STOP_Y;
             Udp::pThis->reply.val = currentValue;
@@ -207,13 +207,13 @@ void StepX::fast_slow(Dir dir) {
 
 void StepX::handler() {
     // currentValue = (pwm.counterX) / coeff;
-    currentValue = (pwm.counter) / coeff;
+    currentValue = (uint32_t)((float)(pwm.counter) / coeff);
     switch (currentState) {
     case STOP:
         break;
     case START_MOVING:
         // if (currentValue < stopValue) {
-        if (pwm.counter < stopValue * coeff) {
+        if (pwm.counter < (uint32_t)((float)stopValue * coeff)) {
             currentState = MOVING_PLUS;
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::START_X;
@@ -222,7 +222,7 @@ void StepX::handler() {
             startPlus();
             // TODO: block all
             //} else if (currentValue > stopValue) {
-        } else if (pwm.counter > stopValue * coeff) {
+        } else if (pwm.counter > (uint32_t)((float)stopValue * coeff)) {
             currentState = MOVING_MINUS;
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::START_X;
@@ -239,11 +239,11 @@ void StepX::handler() {
         }
         break;
     case MOVING_PLUS:
-        if (pwm.counter >= stopValue * coeff) {
+        if (pwm.counter >= (uint32_t)((float)stopValue * coeff)) {
             stop();
             currentState = STOP;
             // TODO: send signal about final place
-            currentValue = (pwm.counter) / coeff;
+            currentValue = (uint32_t)((float)(pwm.counter) / coeff);
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::STOP_X;
             Udp::pThis->reply.val = currentValue;
@@ -277,12 +277,12 @@ void StepX::handler() {
         break;
     case MOVING_MINUS:
         // if (currentValue <= stopValue) {
-        if (pwm.counter <= stopValue * coeff) {
+        if (pwm.counter <= (uint32_t)((float)stopValue * coeff)) {
             // if (pwm.counterX * coeff >= stopValue) {
             stop();
             currentState = STOP;
             // TODO: send signal about final place
-            currentValue = (pwm.counter) / coeff;
+            currentValue = (uint32_t)((float)(pwm.counter) / coeff);
             Udp::pThis->MustSend = true;
             Udp::pThis->reply.currentReply = Protocol::STOP_X;
             Udp::pThis->reply.val = currentValue;
